PES-88: carriers loading update

+++ packetery/packetery.php
@@ -933,70 +933,62 @@ END;
         $key = Configuration::get('PACKETERY_API_KEY');
         $files = array(
             _PS_MODULE_DIR_ . "packetery/address-delivery.xml" =>
-                "https://www.zasilkovna.cz/api/v4/$key/branch.xml"
+                "https://www.zasilkovna.cz/api/v4/$key/branch.xml?address-delivery"
         );
 
         foreach ($files as $local => $remote) {
             if (file_exists($local) && filesize($local) === 0) {
                 unlink($local);
             }
-            if (date("d.m.Y", @filemtime($local)) != date("d.m.Y") && (!file_exists($local) || date("H") >= 1)) {
+            if (!file_exists($local) || date('Y-m-d', filemtime($local)) !== date('Y-m-d')) {
                 if ($this->configuration_errors()) {
-                    if (file_exists($local)) {
-                        $error_count = @Tools::file_get_contents($local . ".error");
-                        if ($error_count > 5) {
-                            unlink($local);
-                        } else {
-                            touch($local);
-                        }
-                        @file_put_contents($local . ".error", $error_count + 1);
+                    if (file_exists($local) && time() - filemtime($local) > 5 * 24 * 60 * 60) {
+                        unlink($local);
                     }
+
                     return;
                 }
-
                 $data = $this->parseBranches($remote);
-
-                file_put_contents($local, $data);
-                @unlink($local . ".error");
+                if ($data) {
+                    file_put_contents($local, $data);
+                }
             }
         }
     }
 
     /**
      * Parses through carriers list and selects address deliveries only
-     * @param $branch_url
-     * @return bool|mixed
+     * @param string $branch_url
+     * @return string|bool
      */
     public function parseBranches($branch_url)
     {
         ignore_user_abort(true);
-        $module = new Packetery();
-        if ($response = Tools::file_get_contents($branch_url)) {
-            if (Tools::strpos($response, 'invalid API key') == false) {
-                $xml = simplexml_load_string($response);
 
-                $toRemove = [];
-                foreach ($xml->carriers->carrier as $k => $carrier) {
-                    if ("$carrier->pickupPoints" == "true") {
-                        $dom = dom_import_simplexml($carrier);
-                        $toRemove[] = $dom;
-                    }
-                }
+        $response = Tools::file_get_contents($branch_url);
+        if (!$response) {
 
-                $dom = dom_import_simplexml($xml->branches);
-                $toRemove[] = $dom;
-
-                foreach ($toRemove as $row) {
-                    $row->parentNode->removeChild($row);
-                }
-
-                return $xml->asXML();
-            } else {
-                return false;
-            }
-        } else {
             return false;
         }
+
+        $xml = simplexml_load_string($response);
+        if ($xml === false) {
+
+            return false;
+        }
+
+        $toRemove = [];
+        foreach ($xml->carriers->carrier as $carrier) {
+            if ((string)$carrier->pickupPoints === "true") {
+                $dom = dom_import_simplexml($carrier);
+                $toRemove[] = $dom;
+            }
+        }
+        foreach ($toRemove as $row) {
+            $row->parentNode->removeChild($row);
+        }
+
+        return $xml->asXML();
     }
 
     /**
